+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - Issue [#493](https://github.com/m-marini/wheellyj/issues/493): Add relative radar map
+- Issue [#496](https://github.com/m-marini/wheellyj/issues/496): Add map relative direction signal
+
 
 ## [0.20.0] 2024-12-28